moved code to XPlatUtility
